update location dto and GoodsService

+++ src/main/java/com/airgear/exception/LocationException.java
@@ -5,7 +5,7 @@ import org.springframework.web.server.ResponseStatusException;
 
 public class LocationException {
 
-    public static ResponseStatusException locationNotFound(Long locationId) {
+    public static ResponseStatusException locationNotFound(Integer locationId) {
         throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Location with id " + locationId + " not found");
     }
